Refactored some stuff
im tired its 2am someone help
